メンテナンス性の向上のため認証ヘッダーの取得等のコードをauth.tsに集約
